Extract ConfigurationCacheConventionMappingIntegrationTest
from ConfigurationCacheIntegrationTest
