+++ subprojects/configuration-cache/src/integTest/groovy/org/gradle/configurationcache/ConfigurationCacheIntegrationTest.groovy
@@ -19,15 +19,9 @@ package org.gradle.configurationcache
 import com.google.common.collect.ImmutableList
 import com.google.common.collect.ImmutableMap
 import com.google.common.collect.ImmutableSet
-import org.gradle.api.Plugin
-import org.gradle.api.Project
 import org.gradle.api.file.ArchiveOperations
 import org.gradle.api.file.FileSystemOperations
-import org.gradle.api.internal.ConventionTask
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.Optional
-import org.gradle.api.tasks.TaskAction
 import org.gradle.initialization.LoadProjectsBuildOperationType
 import org.gradle.initialization.StartParameterBuildOptions.ConfigurationCacheRecreateOption
 import org.gradle.integtests.fixtures.BuildOperationsFixture
@@ -341,57 +335,6 @@ class ConfigurationCacheIntegrationTest extends AbstractConfigurationCacheIntegr
         outputContains("bean.value = child")
         outputContains("bean.parent.value = parent")
         outputContains("same reference = true")
-    }
-
-    def "restores convention mapped task input property explicitly set to null"() {
-        given:
-        withConventionMappingForPropertyOfType String, '"42"'
-        buildFile << '''
-            tasks.named("ok") {
-                inputProperty = null
-            }
-        '''
-
-        when:
-        configurationCacheRun 'ok'
-        configurationCacheRun 'ok'
-
-        then:
-        outputContains 'this.value = null'
-    }
-
-    @Unroll
-    def "restores convention mapped task input property named after field with value of type #typeName"() {
-        given:
-        withConventionMappingForPropertyOfType type, value
-
-        when:
-        configurationCacheRun "ok"
-        configurationCacheRun "ok"
-
-        then:
-        outputContains("this.value = ${output}")
-
-        where:
-        type      | value     | output
-        String    | "'value'" | "value"
-        Boolean   | "true"    | "true"
-        boolean   | "true"    | "true"
-        Character | "'a'"     | "a"
-        char      | "'a'"     | "a"
-        Byte      | "12"      | "12"
-//        byte| "12"      | "12" // TODO: currently not working
-        Short     | "12"      | "12"
-        short     | "12"      | "12"
-        Integer   | "12"      | "12"
-        int       | "12"      | "12"
-        Long      | "12"      | "12"
-        long      | "12"      | "12"
-        Float     | "12.1"    | "12.1"
-        float     | "12.1"    | "12.1"
-        Double    | "12.1"    | "12.1"
-        double    | "12.1"    | "12.1"
-        typeName = type.name
     }
 
     @Unroll
@@ -1214,48 +1157,5 @@ class ConfigurationCacheIntegrationTest extends AbstractConfigurationCacheIntegr
 
         then:
         outputContains("value = value")
-    }
-
-    void withConventionMappingForPropertyOfType(Class type, String value) {
-        final String typeName = type.name
-        file('buildSrc/src/main/java/my/ConventionPlugin.java') << """
-            package my;
-            public class ConventionPlugin implements ${Plugin.name}<${Project.name}> {
-                @Override
-                public void apply(${Project.name} project) {
-                    final Extension ext = project.getExtensions().create("conventions", Extension.class);
-                    project.getTasks().withType(SomeTask.class).configureEach(task -> {
-                        task.getConventionMapping().map("inputProperty", ext::getInputProperty);
-                    });
-                    project.getTasks().register("ok", SomeTask.class, task -> {
-                    });
-                }
-
-                public static abstract class Extension {
-                    private $typeName value;
-                    public $typeName getInputProperty() { return value; }
-                    public void setInputProperty($typeName value) { this.value = value; }
-                }
-
-                public static abstract class SomeTask extends ${ConventionTask.name} {
-                    // Configuration cache only supports convention mapping for fields with matching names.
-                    private $typeName inputProperty;
-                    ${type.primitive ? '' : "@${Optional.name}"}
-                    @${Input.name}
-                    public $typeName getInputProperty() { return inputProperty; }
-                    public void setInputProperty($typeName value) { this.inputProperty = value; }
-                    @${TaskAction.name}
-                    void run() {
-                        System.out.println("this.value = " + getInputProperty());
-                    }
-                }
-            }
-        """
-        buildFile """
-            apply plugin: my.ConventionPlugin
-            conventions {
-                inputProperty = $value
-            }
-        """
     }
 }
